Refactor util into its own folder
util.py being ~500 LOC was becoming hard to maintain. Also, it had some
method tied to real TensorFlow whereas some did not. Now, we broke the
module down into smaller size.
