+++ tensorboard/loader_test.py
@@ -25,8 +25,8 @@ import tensorflow as tf
 
 from tensorboard import db
 from tensorboard import loader
-from tensorboard import test_util
-from tensorboard import util
+from tensorboard.util import test_util
+from tensorboard.util import util
 
 
 class LoaderTestCase(test_util.TestCase):
