+++ tensorboard/plugins/image/BUILD
@@ -65,7 +65,7 @@ py_library(
     deps = [
         ":metadata",
         "//tensorboard:expect_tensorflow_installed",
-        "//tensorboard:util",
+        "//tensorboard/util:encoder",
     ],
 )
