+++ tensorboard/plugins/text/BUILD
@@ -66,7 +66,6 @@ py_library(
     deps = [
         ":metadata",
         "//tensorboard:expect_tensorflow_installed",
-        "//tensorboard:util",
     ],
 )
