+++ tensorboard/util/util.py
@@ -29,10 +29,8 @@ import logging
 import os
 import re
 import sys
-import threading
 import time
 
-import numpy as np
 import six
 
 from tensorboard.compat import tf
@@ -371,162 +369,3 @@ def _hack_the_main_frame():
       'tf_logging.py' in frame.f_back.f_code.co_filename):
     return frame.f_back
   return frame
-
-
-class PersistentOpEvaluator(object):
-  """Evaluate a fixed TensorFlow graph repeatedly, safely, efficiently.
-
-  Extend this class to create a particular kind of op evaluator, like an
-  image encoder. In `initialize_graph`, create an appropriate TensorFlow
-  graph with placeholder inputs. In `run`, evaluate this graph and
-  return its result. This class will manage a singleton graph and
-  session to preserve memory usage, and will ensure that this graph and
-  session do not interfere with other concurrent sessions.
-
-  A subclass of this class offers a threadsafe, highly parallel Python
-  entry point for evaluating a particular TensorFlow graph.
-
-  Example usage:
-
-      class FluxCapacitanceEvaluator(PersistentOpEvaluator):
-        \"\"\"Compute the flux capacitance required for a system.
-
-        Arguments:
-          x: Available power input, as a `float`, in jigawatts.
-
-        Returns:
-          A `float`, in nanofarads.
-        \"\"\"
-
-        def initialize_graph(self):
-          self._placeholder = tf.placeholder(some_dtype)
-          self._op = some_op(self._placeholder)
-
-        def run(self, x):
-          return self._op.eval(feed_dict: {self._placeholder: x})
-
-      evaluate_flux_capacitance = FluxCapacitanceEvaluator()
-
-      for x in xs:
-        evaluate_flux_capacitance(x)
-  """
-
-  def __init__(self):
-    super(PersistentOpEvaluator, self).__init__()
-    self._session = None
-    self._initialization_lock = threading.Lock()
-
-  def _lazily_initialize(self):
-    """Initialize the graph and session, if this has not yet been done."""
-    with self._initialization_lock:
-      if self._session:
-        return
-      graph = tf.Graph()
-      with graph.as_default():
-        self.initialize_graph()
-      # Don't reserve GPU because libpng can't run on GPU.
-      config = tf.ConfigProto(device_count={'GPU': 0})
-      self._session = tf.Session(graph=graph, config=config)
-
-  def initialize_graph(self):
-    """Create the TensorFlow graph needed to compute this operation.
-
-    This should write ops to the default graph and return `None`.
-    """
-    raise NotImplementedError('Subclasses must implement "initialize_graph".')
-
-  def run(self, *args, **kwargs):
-    """Evaluate the ops with the given input.
-
-    When this function is called, the default session will have the
-    graph defined by a previous call to `initialize_graph`. This
-    function should evaluate any ops necessary to compute the result of
-    the query for the given *args and **kwargs, likely returning the
-    result of a call to `some_op.eval(...)`.
-    """
-    raise NotImplementedError('Subclasses must implement "run".')
-
-  def __call__(self, *args, **kwargs):
-    self._lazily_initialize()
-    with self._session.as_default():
-      return self.run(*args, **kwargs)
-
-
-class _TensorFlowPngEncoder(PersistentOpEvaluator):
-  """Encode an image to PNG.
-
-  This function is thread-safe, and has high performance when run in
-  parallel. See `encode_png_benchmark.py` for details.
-
-  Arguments:
-    image: A numpy array of shape `[height, width, channels]`, where
-      `channels` is 1, 3, or 4, and of dtype uint8.
-
-  Returns:
-    A bytestring with PNG-encoded data.
-  """
-
-  def __init__(self):
-    super(_TensorFlowPngEncoder, self).__init__()
-    self._image_placeholder = None
-    self._encode_op = None
-
-  def initialize_graph(self):
-    self._image_placeholder = tf.placeholder(
-        dtype=tf.uint8, name='image_to_encode')
-    self._encode_op = tf.image.encode_png(self._image_placeholder)
-
-  def run(self, image):  # pylint: disable=arguments-differ
-    if not isinstance(image, np.ndarray):
-      raise ValueError("'image' must be a numpy array: %r" % image)
-    if image.dtype != np.uint8:
-      raise ValueError("'image' dtype must be uint8, but is %r" % image.dtype)
-    return self._encode_op.eval(feed_dict={self._image_placeholder: image})
-
-
-encode_png = _TensorFlowPngEncoder()
-
-
-class _TensorFlowWavEncoder(PersistentOpEvaluator):
-  """Encode an audio clip to WAV.
-
-  This function is thread-safe and exhibits good parallel performance.
-
-  Arguments:
-    audio: A numpy array of shape `[samples, channels]`.
-    samples_per_second: A positive `int`, in Hz.
-
-  Returns:
-    A bytestring with WAV-encoded data.
-  """
-
-  def __init__(self):
-    super(_TensorFlowWavEncoder, self).__init__()
-    self._audio_placeholder = None
-    self._samples_per_second_placeholder = None
-    self._encode_op = None
-
-  def initialize_graph(self):
-    self._audio_placeholder = tf.placeholder(
-        dtype=tf.float32, name='image_to_encode')
-    self._samples_per_second_placeholder = tf.placeholder(
-        dtype=tf.int32, name='samples_per_second')
-    self._encode_op = tf.contrib.ffmpeg.encode_audio(
-        self._audio_placeholder,
-        file_format='wav',
-        samples_per_second=self._samples_per_second_placeholder)
-
-  def run(self, audio, samples_per_second):  # pylint: disable=arguments-differ
-    if not isinstance(audio, np.ndarray):
-      raise ValueError("'audio' must be a numpy array: %r" % audio)
-    if not isinstance(samples_per_second, int):
-      raise ValueError("'samples_per_second' must be an int: %r"
-                       % samples_per_second)
-    feed_dict = {
-        self._audio_placeholder: audio,
-        self._samples_per_second_placeholder: samples_per_second,
-    }
-    return self._encode_op.eval(feed_dict=feed_dict)
-
-
-encode_wav = _TensorFlowWavEncoder()
