+++ tensorboard/util/util_test.py
@@ -21,11 +21,10 @@ import logging
 import time
 
 import six
-import numpy as np
 import tensorflow as tf
 
-from tensorboard import test_util
-from tensorboard import util
+from tensorboard.util import test_util
+from tensorboard.util import util
 
 
 class LogFormatterTest(tf.test.TestCase):
@@ -226,108 +225,6 @@ def TerminalStringIO():
   stream = six.StringIO()
   stream.isatty = lambda: True
   return stream
-
-
-class PersistentOpEvaluatorTest(tf.test.TestCase):
-
-  def setUp(self):
-    super(PersistentOpEvaluatorTest, self).setUp()
-
-    patch = tf.test.mock.patch('tensorflow.Session', wraps=tf.Session)
-    patch.start()
-    self.addCleanup(patch.stop)
-
-    class Squarer(util.PersistentOpEvaluator):
-
-      def __init__(self):
-        super(Squarer, self).__init__()
-        self._input = None
-        self._squarer = None
-
-      def initialize_graph(self):
-        self._input = tf.placeholder(tf.int32)
-        self._squarer = tf.square(self._input)
-
-      def run(self, xs):  # pylint: disable=arguments-differ
-        return self._squarer.eval(feed_dict={self._input: xs})
-
-    self._square = Squarer()
-
-  def test_preserves_existing_session(self):
-    with tf.Session() as sess:
-      op = tf.reduce_sum([2, 2])
-      self.assertIs(sess, tf.get_default_session())
-
-      result = self._square(123)
-      self.assertEqual(123 * 123, result)
-
-      self.assertIs(sess, tf.get_default_session())
-      number_of_lights = sess.run(op)
-      self.assertEqual(number_of_lights, 4)
-
-  def test_lazily_initializes_sessions(self):
-    self.assertEqual(tf.Session.call_count, 0)
-
-  def test_reuses_sessions(self):
-    self._square(123)
-    self.assertEqual(tf.Session.call_count, 1)
-    self._square(234)
-    self.assertEqual(tf.Session.call_count, 1)
-
-
-class TensorFlowPngEncoderTest(tf.test.TestCase):
-
-  def setUp(self):
-    super(TensorFlowPngEncoderTest, self).setUp()
-    self._encode = util._TensorFlowPngEncoder()
-    self._rgb = np.arange(12 * 34 * 3).reshape((12, 34, 3)).astype(np.uint8)
-    self._rgba = np.arange(21 * 43 * 4).reshape((21, 43, 4)).astype(np.uint8)
-
-  def _check_png(self, data):
-    # If it has a valid PNG header and is of a reasonable size, we can
-    # assume it did the right thing. We trust the underlying
-    # `encode_png` op.
-    self.assertEqual(b'\x89PNG', data[:4])
-    self.assertGreater(len(data), 128)
-
-  def test_invalid_non_numpy(self):
-    with six.assertRaisesRegex(self, ValueError, "must be a numpy array"):
-      self._encode(self._rgb.tolist())
-
-  def test_invalid_non_uint8(self):
-    with six.assertRaisesRegex(self, ValueError, "dtype must be uint8"):
-      self._encode(self._rgb.astype(np.float32))
-
-  def test_encodes_png(self):
-    data = self._encode(self._rgb)
-    self._check_png(data)
-
-  def test_encodes_png_with_alpha(self):
-    data = self._encode(self._rgba)
-    self._check_png(data)
-
-
-class TensorFlowWavEncoderTest(tf.test.TestCase):
-
-  def setUp(self):
-    super(TensorFlowWavEncoderTest, self).setUp()
-    self._encode = util._TensorFlowWavEncoder()
-    space = np.linspace(0.0, 100.0, 44100)
-    self._stereo = np.array([np.sin(space), np.cos(space)]).transpose()
-    self._mono = self._stereo.mean(axis=1, keepdims=True)
-
-  def _check_wav(self, data):
-    # If it has a valid WAV/RIFF header and is of a reasonable size, we
-    # can assume it did the right thing. We trust the underlying
-    # `encode_audio` op.
-    self.assertEqual(b'RIFF', data[:4])
-    self.assertGreater(len(data), 128)
-
-  def test_encodes_mono_wav(self):
-    self._check_wav(self._encode(self._mono, samples_per_second=44100))
-
-  def test_encodes_stereo_wav(self):
-    self._check_wav(self._encode(self._stereo, samples_per_second=44100))
 
 
 if __name__ == '__main__':
